api/requirements.txt
asgiref==3.7.2
Django==5.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
